[Trunk] - Small bug in DbaseReader removed (TypeByTypeCode returned int instead of long for Int64 -> Failed to read shapefiles with Int64 columns - Small bug in MapBox removed (Extending Bounding Box for empty layers) - Rainbow7 ColorBlend fixed
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS01]$/SharpMap/Trunk;C100446

+++ SharpMap/Data/Providers/DbaseReader.cs
@@ -727,7 +727,7 @@ namespace SharpMap.Data.Providers
                 case TypeCode.Int32:
                     return typeof (int);
                 case TypeCode.Int64:
-                    return typeof (int);
+                    return typeof (long);
                 case TypeCode.Single:
                     return typeof (float);
                 case TypeCode.Double:
